simplify fsdp2 config to reduce version complexity first
